change `uint` to `size_t` (thanks @DennisMitchel)

+++ ecndpcaalrlp.d
@@ -41,7 +41,7 @@ class EcndpcaalrlpState {
         code = input;
         push(0);
         // parse jumpPositions
-        uint[] positionStack;
+        size_t[] positionStack;
         foreach(i, c; code) {
             if(c == '(') {
                 positionStack ~= i;
